No multiple snips allowed.
I'm trying to keep things simple. From now on, assumption is, that if
someone define a snippet for global or certain filetype, which already
exists in external repository (like in ~/.vim/bundle/…/snippets), it
will be taken first, and the one in bundle path ignored.

This is using the fact, that runtimepath starts from ~/.vimrc, so the
place where the user configuration lies. Whatever user is defined takes
precedence.